feat(computer): add Linux, macOS, and Windows support for desktop automation; refactor port handling utilities
