+++ apps/site/rspress.config.ts
@@ -135,21 +135,21 @@ export default defineConfig({
           text: 'API reference (iOS)',
           link: '/ios-api-reference',
         },
-        {
-          sectionHeaderText: 'PC Desktop',
-        },
-        {
-          text: 'Introduction',
-          link: '/computer-introduction',
-        },
-        {
-          text: 'Getting started',
-          link: '/computer-getting-started',
-        },
-        {
-          text: 'API reference (PC Desktop)',
-          link: '/computer-api-reference',
-        },
+        // {
+        //   sectionHeaderText: 'PC Desktop',
+        // },
+        // {
+        //   text: 'Introduction',
+        //   link: '/computer-introduction',
+        // },
+        // {
+        //   text: 'Getting started',
+        //   link: '/computer-getting-started',
+        // },
+        // {
+        //   text: 'API reference (PC Desktop)',
+        //   link: '/computer-api-reference',
+        // },
         {
           sectionHeaderText: 'More features',
         },
@@ -289,21 +289,21 @@ export default defineConfig({
           text: 'API 参考（iOS）',
           link: '/zh/ios-api-reference',
         },
-        {
-          sectionHeaderText: 'PC 桌面',
-        },
-        {
-          text: '介绍',
-          link: '/zh/computer-introduction',
-        },
-        {
-          text: '开始使用',
-          link: '/zh/computer-getting-started',
-        },
-        {
-          text: 'API 参考（PC 桌面）',
-          link: '/zh/computer-api-reference',
-        },
+        // {
+        //   sectionHeaderText: 'PC 桌面',
+        // },
+        // {
+        //   text: '介绍',
+        //   link: '/zh/computer-introduction',
+        // },
+        // {
+        //   text: '开始使用',
+        //   link: '/zh/computer-getting-started',
+        // },
+        // {
+        //   text: 'API 参考（PC 桌面）',
+        //   link: '/zh/computer-api-reference',
+        // },
         {
           sectionHeaderText: '更多特性',
         },
