+++ packages/android-playground/src/bin.ts
@@ -1,5 +1,4 @@
 import { exec } from 'node:child_process';
-import { createServer } from 'node:net';
 import path from 'node:path';
 import { promisify } from 'node:util';
 import { select } from '@inquirer/prompts';
@@ -13,37 +12,10 @@ import {
   PLAYGROUND_SERVER_PORT,
   SCRCPY_SERVER_PORT,
 } from '@midscene/shared/constants';
+import { findAvailablePort } from '@midscene/shared/node';
 import ScrcpyServer from './scrcpy-server';
 
 const promiseExec = promisify(exec);
-
-async function isPortAvailable(port: number): Promise<boolean> {
-  return new Promise((resolve) => {
-    const server = createServer();
-    server.on('error', () => resolve(false));
-    server.listen(port, () => {
-      server.close(() => resolve(true));
-    });
-  });
-}
-
-async function findAvailablePort(startPort: number): Promise<number> {
-  let port = startPort;
-  let attempts = 0;
-  const maxAttempts = 15;
-
-  while (!(await isPortAvailable(port))) {
-    attempts++;
-    if (attempts >= maxAttempts) {
-      console.error(
-        `❌ Unable to find available port after ${maxAttempts} attempts starting from ${startPort}`,
-      );
-      process.exit(1);
-    }
-    port++;
-  }
-  return port;
-}
 
 // Function to get available devices
 async function getAdbDevices() {
