+++ packages/computer-linux/src/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Linux-specific re-export of @midscene/computer
+ * This package provides the same functionality as @midscene/computer
+ * but is specifically for Linux platform.
+ */
+export * from '@midscene/computer';
