+++ packages/computer-mac/rslib.config.ts
@@ -1,0 +1,25 @@
+import { defineConfig } from '@rslib/core';
+
+export default defineConfig({
+  lib: [
+    {
+      output: { distPath: { root: 'dist/lib' } },
+      format: 'cjs',
+      syntax: 'es2020',
+    },
+    {
+      output: { distPath: { root: 'dist/es' } },
+      dts: {
+        bundle: true,
+        distPath: 'dist/types',
+      },
+      format: 'esm',
+      syntax: 'es2020',
+    },
+  ],
+  source: {
+    entry: {
+      index: './src/index.ts',
+    },
+  },
+});
