+++ packages/computer-mac/src/index.ts
@@ -1,0 +1,6 @@
+/**
+ * macOS-specific re-export of @midscene/computer
+ * This package provides the same functionality as @midscene/computer
+ * but is specifically for macOS platform.
+ */
+export * from '@midscene/computer';
