+++ packages/computer-mac/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "extends": "../shared/tsconfig.base.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "module": "ES2020",
+    "declarationDir": "dist/types",
+    "types": ["node"]
+  },
+  "include": ["src"],
+  "references": [
+    {
+      "path": "../computer"
+    }
+  ]
+}
