+++ packages/computer-playground/src/bin.ts
@@ -1,38 +1,10 @@
-import { createServer } from 'node:net';
 import os from 'node:os';
 import path from 'node:path';
 import { ComputerDevice, agentFromComputer } from '@midscene/computer';
 import { PlaygroundServer } from '@midscene/playground';
 import { PLAYGROUND_SERVER_PORT } from '@midscene/shared/constants';
+import { findAvailablePort } from '@midscene/shared/node';
 import puppeteer from 'puppeteer';
-
-async function isPortAvailable(port: number): Promise<boolean> {
-  return new Promise((resolve) => {
-    const server = createServer();
-    server.on('error', () => resolve(false));
-    server.listen(port, () => {
-      server.close(() => resolve(true));
-    });
-  });
-}
-
-async function findAvailablePort(startPort: number): Promise<number> {
-  let port = startPort;
-  let attempts = 0;
-  const maxAttempts = 15;
-
-  while (!(await isPortAvailable(port))) {
-    attempts++;
-    if (attempts >= maxAttempts) {
-      console.error(
-        `❌ Unable to find available port after ${maxAttempts} attempts starting from ${startPort}`,
-      );
-      process.exit(1);
-    }
-    port++;
-  }
-  return port;
-}
 
 const staticDir = path.join(__dirname, '../../static');
