+++ packages/shared/src/node/index.ts
@@ -1,1 +1,2 @@
 export * from './fs';
+export * from './port';
